Error handling in data load

+++ data_load/dataLoad.py
@@ -1,6 +1,7 @@
 import csv
 import sys
 import json
+import os.path
 from pymongo import MongoClient
 
 def initCollection(collectionName, db):
@@ -32,18 +33,22 @@ else:
     mClient = MongoClient()
     mDb = mClient.tstorydb
     for fileName in sys.argv[1:len(sys.argv)]:
-        collectionName = fileName.split('/')
-        collectionName = collectionName[len(collectionName)-1].split('.')[0]
-        dropCollection(collectionName, mDb)
-        initCollection(collectionName, mDb)
-        with open(fileName,'rb') as csvFile:
-            docArray = []
-            reader = csv.reader(csvFile, quotechar='"')
-            next(reader, None)
-            for row in reader:
-                doc = createDocument(row)
-                print("Adding document = " + json.dumps(doc))
-                docArray.append(doc)
-        if collectionName in mDb.list_collection_names():
-            mDb[collectionName].insert_many(docArray)
-            print("Data loaded into " + collectionName)
+        if os.path.isfile(fileName):
+            collectionName = fileName.split('/')
+            collectionName = collectionName[len(collectionName)-1].split('.')[0]
+            dropCollection(collectionName, mDb)
+            initCollection(collectionName, mDb)
+            with open(fileName,'rb') as csvFile:
+                docArray = []
+                reader = csv.reader(csvFile, quotechar='"')
+                next(reader, None)
+                for row in reader:
+                    doc = createDocument(row)
+                    print("Adding document = " + json.dumps(doc))
+                    docArray.append(doc)
+            if collectionName in mDb.list_collection_names():
+                mDb[collectionName].insert_many(docArray)
+                print("Data loaded into " + collectionName)
+        else:
+            print(fileName + " does not exist!")
+            sys.exit(0)
